Move configurations to the plugin
