+++ Modules/Sources/BuildSettings/BuildSettings.swift
@@ -43,25 +43,3 @@ public struct BuildSettings: Sendable, Codable {
 //        _shared = settings
     }
 }
-
-public enum BuildTarget: String {
-    /// WordPress app.
-    case wordpress
-
-    /// Jetpack app.
-    case jetpack
-}
-
-public enum BuildConfiguration: String {
-    /// Development build, usually run from Xcode
-    case localDeveloper
-
-    /// Continuous integration builds for Automattic employees to test branches & PRs
-    case a8cBranchTest
-
-    /// Beta released internally for Automattic employees
-    case a8cPrereleaseTesting
-
-    /// Production build released in the app store
-    case appStore
-}
