--- Targets/Jetpack/Sources/AppDelegate+Jetpack.swift
@@ -1,7 +1,0 @@
-import UIKit
-
-class AppDelegate: WordPressAppDelegate {
-    override func makeBuildSettings() -> BuildSettings {
-        BuildSettings.make()
-    }
-}
