--- Targets/Jetpack/Sources/BuildSettings+Jetpack.swift
@@ -1,4 +1,0 @@
-import UIKit
-import BuildSettings
-
-
