+++ WordPress/Classes/System/main.swift
@@ -1,9 +1,7 @@
-import Foundation
 import UIKit
-import BuildSettings
 
 let isRunningTests = NSClassFromString("XCTestCase") != nil
-let appDelegateClass = isRunningTests ? "TestingAppDelegate" : NSStringFromClass(AppDelegate.self)
+let appDelegateClass = isRunningTests ? "TestingAppDelegate" : NSStringFromClass(WordPressAppDelegate.self)
 
 UIApplicationMain(
     CommandLine.argc,
